+++ src/codemirror.ts
@@ -27,14 +27,14 @@ import {mediawiki, html} from './mediawiki';
 import {escapeKeymap} from './escape';
 import {foldExtension, foldHandler} from './fold';
 import {tagMatchingState} from './matchTag';
-import {CDN, loadScript} from './util';
+import {CDN} from './util';
+import {getWikiLinter, getJsLinter, getCssLinter, getLuaLinter, getJsonLinter} from './linter';
 import * as plugins from './plugins';
 import type {ViewPlugin, KeyBinding} from '@codemirror/view';
 import type {Extension, Text, StateEffect} from '@codemirror/state';
 import type {SyntaxNode} from '@lezer/common';
 import type {Diagnostic, Action} from '@codemirror/lint';
 import type {Highlighter} from '@lezer/highlight';
-import type {Linter} from 'eslint';
 import type {Config} from 'wikiparser-node';
 import type {MwConfig} from './mediawiki';
 import type {DocRange} from './fold';
@@ -330,37 +330,12 @@ export class CodeMirror6 {
 	async getLinter(opt?: Record<string, unknown>): Promise<LintSource | undefined> {
 		switch (this.#lang) {
 			case 'mediawiki': {
-				const REPO = 'npm/wikiparser-node@1.6.2-b',
-					DIR = `${REPO}/extensions/dist`,
-					src = `combine/${DIR}/base.min.js,${DIR}/lint.min.js`,
-					lang = opt?.['i18n'];
-				await loadScript(src, 'wikiparse');
-				if (typeof lang === 'string') {
-					try {
-						const i18n: Record<string, string>
-							= await (await fetch(`${CDN}/${REPO}/i18n/${lang.toLowerCase()}.json`)).json();
-						wikiparse.setI18N(i18n);
-					} catch {}
-				}
-				const wikiLinter = new wikiparse.Linter!(opt?.['include'] as boolean | undefined);
+				const wikiLinter = await getWikiLinter(opt);
 				return doc => wikiLinter.codemirror(doc.toString());
 			}
 			case 'javascript': {
-				await loadScript('npm/eslint-linter-browserify', 'eslint');
-				/** @see https://www.npmjs.com/package/@codemirror/lang-javascript */
-				const esLinter = new eslint.Linter(),
-					conf: Linter.Config = {
-						env: {browser: true, es2024: true},
-						parserOptions: {ecmaVersion: 15, sourceType: 'module'},
-						rules: {},
-						...opt,
-					};
-				for (const [name, {meta}] of esLinter.getRules()) {
-					if (meta?.docs?.recommended) {
-						conf.rules![name] ??= 2;
-					}
-				}
-				return doc => esLinter.verify(doc.toString(), conf)
+				const esLint = await getJsLinter(opt);
+				return doc => esLint(doc.toString())
 					.map(({ruleId, message, severity, line, column, endLine, endColumn, fix, suggestions = []}) => {
 						const start = pos(doc, line, column),
 							diagnostic: Diagnostic = {
@@ -385,121 +360,44 @@ export class CodeMirror6 {
 					});
 			}
 			case 'css': {
-				await loadScript('gh/openstyles/stylelint-bundle/dist/stylelint-bundle.min.js', 'stylelint');
-				/** @see https://www.npmjs.com/package/stylelint-config-recommended */
-				const config = {
-					rules: {
-						'annotation-no-unknown': true,
-						'at-rule-no-unknown': true,
-						'block-no-empty': true,
-						'color-no-invalid-hex': true,
-						'comment-no-empty': true,
-						'custom-property-no-missing-var-function': true,
-						'declaration-block-no-duplicate-custom-properties': true,
-						'declaration-block-no-duplicate-properties': [
-							true,
+				const styleLint = await getCssLinter(opt);
+				return async doc => (await styleLint(doc.toString()))
+					.map(({text, severity, line, column, endLine, endColumn}) => ({
+						source: 'Stylelint',
+						message: text,
+						severity,
+						from: pos(doc, line, column),
+						to: endLine === undefined ? doc.line(line).to : pos(doc, endLine, endColumn!),
+					}));
+			}
+			case 'lua': {
+				const luaLint = await getLuaLinter();
+				return doc => luaLint(doc.toString());
+			}
+			case 'json': {
+				const jsonLint = getJsonLinter();
+				return doc => {
+					const [e] = jsonLint(doc.toString());
+					if (e) {
+						const {message, line, column, position} = e;
+						let from = 0;
+						if (position) {
+							from = Number(position);
+						} else if (line && column) {
+							from = pos(doc, Number(line), Number(column));
+						}
+						return [
 							{
-								ignore: ['consecutive-duplicates-with-different-syntaxes'],
+								message,
+								severity: 'error',
+								from,
+								to: from,
 							},
-						],
-						'declaration-block-no-shorthand-property-overrides': true,
-						'font-family-no-duplicate-names': true,
-						'font-family-no-missing-generic-family-keyword': true,
-						'function-calc-no-unspaced-operator': true,
-						'function-linear-gradient-no-nonstandard-direction': true,
-						'function-no-unknown': true,
-						'keyframe-block-no-duplicate-selectors': true,
-						'keyframe-declaration-no-important': true,
-						'media-feature-name-no-unknown': true,
-						'media-query-no-invalid': true,
-						'named-grid-areas-no-invalid': true,
-						'no-descending-specificity': true,
-						'no-duplicate-at-import-rules': true,
-						'no-duplicate-selectors': true,
-						'no-empty-source': true,
-						'no-invalid-double-slash-comments': true,
-						'no-invalid-position-at-import-rule': true,
-						'no-irregular-whitespace': true,
-						'property-no-unknown': true,
-						'selector-anb-no-unmatchable': true,
-						'selector-pseudo-class-no-unknown': true,
-						'selector-pseudo-element-no-unknown': true,
-						'selector-type-no-unknown': [
-							true,
-							{
-								ignore: ['custom-elements'],
-							},
-						],
-						'string-no-newline': true,
-						'unit-no-unknown': true,
-						...opt?.['rules'] as Record<string, unknown>,
-					},
-				};
-				return async doc => {
-					const {results} = await stylelint.lint({code: doc.toString(), config});
-					return results.flatMap(({warnings}) => warnings)
-						.map(({text, severity, line, column, endLine, endColumn}) => ({
-							source: 'Stylelint',
-							message: text,
-							severity,
-							from: pos(doc, line, column),
-							to: endLine === undefined ? doc.line(line).to : pos(doc, endLine, endColumn!),
-						}));
+						];
+					}
+					return [];
 				};
 			}
-			case 'lua':
-				await loadScript('npm/luaparse', 'luaparse');
-				/** @see https://github.com/ajaxorg/ace/pull/4954 */
-				luaparse.defaultOptions.luaVersion = '5.3';
-				return doc => {
-					try {
-						luaparse.parse(doc.toString());
-					} catch (e) {
-						if (e instanceof luaparse.SyntaxError) {
-							return [
-								{
-									source: 'luaparse',
-									message: e.message.replace(/^\[\d+:\d+\]\s*/u, ''),
-									severity: 'error',
-									from: e.index,
-									to: e.index,
-								},
-							];
-						}
-					}
-					return [];
-				};
-			case 'json':
-				return doc => {
-					try {
-						const str = doc.toString();
-						if (str.trim()) {
-							JSON.parse(str);
-						}
-					} catch (e) {
-						if (e instanceof SyntaxError) {
-							const {message} = e,
-								line = /\bline (\d+)/u.exec(message)?.[1],
-								column = /\bcolumn (\d+)/u.exec(message)?.[1],
-								position = /\bposition (\d+)/u.exec(message)?.[1];
-							let from = 0;
-							if (position) {
-								from = Number(position);
-							} else if (line && column) {
-								from = pos(doc, Number(line), Number(column));
-							}
-							return [
-								{
-									message,
-									severity: 'error',
-									from,
-									to: from,
-								},
-							];
-						}
-					}
-					return [];
-				};
 			default:
 				return undefined;
 		}